Voltando js de portfolio de volta pra blade pq o npm run dev quebrava na hora de copiar o código :/
